// src/2-1-employees/add-employee/types.ts
export interface PersonalDataStepProps {
  formData: any;
  handleInputChange: (field: string, value: any) => void;
}

export interface EmploymentDataStepProps {
  formData: any;
  handleInputChange: (field: string, value: any) => void;
}

export interface InviteStepProps {
  formData: any;
  handleInputChange?: (field: string, value: any) => void;
}